+++ tests/testthat/test-order.R
@@ -137,7 +137,7 @@ test_that("order Test 8: Variable ordering messaging is correct", {
   local_cli_theme()
   suppressMessages(
     xportr_order(df, df_meta, verbose = "message", domain = "df") %>%
-      expect_message("All variables in specification file are in dataset") %>%
+      expect_message("All variables in dataset are found in `metadata`") %>%
       expect_condition("4 reordered in dataset") %>%
       expect_message("Variable reordered in `.df`: `a`, `b`, `c`, and `d`")
   )
@@ -172,13 +172,13 @@ test_that("order Test 10: Gets warning when metadata has multiple rows with same
   # Checks that message appears when xportr.domain_name is invalid
   suppressMessages(multiple_vars_in_spec_helper(xportr_order) %>%
     # expect_message() are being caught to provide clean test without output      #nolint
-    expect_message("All variables in specification file are in dataset") %>%
+    expect_message("All variables in dataset are found in `metadata`") %>%
     expect_message("All variables in dataset are ordered"))
 
   # Checks that message doesn't appear when xportr.domain_name is valid
   suppressMessages(multiple_vars_in_spec_helper2(xportr_order) %>%
     # expect_message() are being caught to provide clean test without output     #nolint
-    expect_message("All variables in specification file are in dataset") %>%
+    expect_message("All variables in dataset are found in `metadata`") %>%
     expect_message("All variables in dataset are ordered"))
 })